requirements.txt
boto3==1.35.71
Django==5.1.3
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.3.3
django-simple-history==3.7.0
django-soft-delete==1.0.15
django-storages==1.14.4
django-tailwind==3.8.0
django-unfold==0.42.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
